pyproject.toml
[project]
name = "scinotebook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bokeh>=3.8.0",
    "catboost>=1.2.8",
    "fastai>=2.8.5",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "jupyter-ai>=2.31.6",
    "jupyterlab-git>=0.51.2",
    "langchain-openai>=0.3.35",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pip>=25.2",
    "plotly>=6.3.1",
    "pymc>=5.26.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "supervision>=0.26.1",
    "sympy>=1.14.0",
    "ultralytics>=8.3.222",
    "xarray>=2025.10.1",
    "yfinance>=0.2.66",
]
